Use vfsStream via composer in development
Instead of assuming vfsStream is on the include path, we now list it as a
development requirement. Additionally, the usage from the original patch was
incorrect, and was creating files relative to the working directory; this patch
updates the usage to correctly setup a vfs root directory and create files
relative to it.