Allow multiple instances of all plugin types
As reported in #27, 2.7.0 removed the override value for the "share by default"
flag in all plugin managers, which breaks backwards compatibility. Previously,
each set the flag to boolean `false`, allowing multiple instances.

This patch adds a test trait that checks for the status of the flag (using the
correct property based on zend-servicemanager release), and composes it into the
plugin manager compatibility tests.