Allow custom ignored namespaces in backtrace processor
The backtrace processor adds the file, line, class, and function
responsible for triggering a log to the event extras.

By default, classes in the backtrace under the Zend\Log namespace are
ignored so that we can see the actual application code that called the
log().

However, there might also be cases when we want to ignore other
namespaces so that we get the true origin. The example I came across
that prompted this patch was Psr\Log, which appears when the
PsrLoggerAdapter wraps the underlying Zend Logger.

This patch allows us to add our own custom ignored namespaces in
addition to the default Zend\Log.