Change the priority value to avoid confused with Priority filter
This priority should not be confused with the Priority filter, which determines if a message meets a severity threshold. When a Priority filter is attached, the writer will output the message only when the filter is lower or equal (by default) to the priority (the biggest severity is 0).
This priority is the priority in the queue of all writers of the logger. It can be any integer, and the default is 1